+++ website/app/docs/Streams/_meta.js
@@ -3,6 +3,8 @@ export default {
   streamToBytes: 'streamToBytes',
   streamToArray: 'streamToArray',
   streamToLines: 'streamToLines',
+  streamToIter: 'streamToIter',
+  iterToStream: 'iterToStream',
   concatStreams: 'concatStreams',
   mergeStreams: 'mergeStreams',
 };